Use Float64 for squall line example
Float32 causes spurious stratospheric cooling from WENO9 truncation
error in this 2D configuration, leading to NaN after ~12 minutes.
Float64 eliminates the issue; full 4-hour run completes cleanly.

+++ examples/squall_line.jl
@@ -78,7 +78,7 @@ using Printf
 # 1 km horizontal resolution and 500 m vertical resolution.
 # The 2D setup uses `(Periodic, Flat, Bounded)` topology.
 
-Oceananigans.defaults.FloatType = Float32
+Oceananigans.defaults.FloatType = Float64
 
 Nx, Nz = 600, 40
 Lx, Lz = 600kilometers, 20kilometers
